+++ src/color/mod.rs
@@ -9,12 +9,14 @@ mod lab;
 mod lch_ab;
 mod lch_uv;
 mod luv;
+mod palette;
 mod rgb;
 mod working_space;
 mod xyy;
 mod xyz;
 
 pub use gradient::Gradient;
+pub use palette::{Palette, PaletteFormat};
 
 pub use chromatic_adaptation::ChromaticAdaptationMethod;
 pub use cmyk::Cmyk;