+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Add an extra black border to the zoom picker window on Linux so that the window is visible on both black and white backgrounds
 - Remove sleep when zoom window is dragged so that the movement in smoother
 - Fix side panel width when no colors are added
+- Add support for multiple palettes
 
 # 0.7.0
 - Fix invisible circle on 2D HSV slider